Pulled WebCode into lua class - compiles now need to test
